pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptforge"
version = "0.1.0"
description = "CLI-first prompt evaluation agent using the OpenAI Responses API"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
  {name = "PromptForge contributors"},
]
dependencies = [
  "jinja2>=3.1.0",
  "jsonschema>=4.0.0",
  "openai>=2.26.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
pf = "promptforge.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
promptforge = ["agents/**/*.md", "agents/**/*.yaml"]

